+++ src/constants/helpers.tsx
@@ -98,7 +98,8 @@ export const localStorageKeys = {
 
 export const formFieldNames = {
   marketSegment: 'marketSegment',
-  subMarketSegment: 'subMarketSegment'
+  subMarketSegment: 'subMarketSegment',
+  parentAccount: 'parentAccount'
 };
 
 export const sidebarResource = {
@@ -383,9 +384,17 @@ export const profileMenuItems = {
 };
 
 export const getObjKeys = (val: string | boolean = '', arr: any[]) => {
+  let selectedEntity = localStorage.getItem("selectedEntity")
+
+  let isCreate = (val === "") ? true : false
   const obj = {};
   for (const key of arr) {
+    let isEntityField = key?.fieldName === "entity"
+
     let value = key.isDefaultValue ? key.defaultValue : val;
+    if (isEntityField && selectedEntity && isCreate) {
+      value = key?.type === "multiSelect" ? [selectedEntity] : selectedEntity
+    }
     if (key.type === 'dropDown') {
       const option = key.option?.find((data: any) => data.default === true);
       obj[key.fieldName] = value ? value : option ? option.optionValue : '';
@@ -434,8 +443,8 @@ export const getObjKeysWithValues = (dataObj: object, arr: any[]) => {
 
   const filterValues = (data: object | any) => (typeof data === 'string' ? data : typeof data === 'object' ? data.optionValue : '');
   for (const key of arr) {
-
     let defaultValue
+
     if (key?.isDefaultValue && key?.defaultValue) {
       defaultValue = key.defaultValue
     }
@@ -443,6 +452,7 @@ export const getObjKeysWithValues = (dataObj: object, arr: any[]) => {
       obj[key.fieldName] = dataObj[key.fieldName] ? dataObj[key.fieldName] : defaultValue ? defaultValue : false;
     } else if (key.type === 'multiSelect') {
       const values = dataObj[key.fieldName] && dataObj[key.fieldName].length
+
         ? typeof dataObj[key.fieldName] === 'string'
           ? dataObj[key.fieldName]
           : dataObj[key.fieldName].map((val: any) => filterValues(val))