+++ src/pages/QuotePdfTemplate/NewCreateQuotePdfTemplate.tsx
@@ -76,7 +76,7 @@ export default function NewCreateQuotePdfTemplate() {
     const [quoteData] = useState(history.location.state?.quoteData);
     const [version] = useState(history.location.state?.version);
     const {
-        state: { user },
+        state: { user, selectedEntity },
     }: any = useData();
     const [ownerCollaboratorData, setOwnerCollaboratorData] = useState([]);
     const [ownerCollaboratorDataConst, setOwnerCollaboratorDataConst] = useState([]);
@@ -184,7 +184,7 @@ export default function NewCreateQuotePdfTemplate() {
                 footer: "",
                 aboveTable: "",
                 belowTable: "",
-                entity: [],
+                entity: selectedEntity ? [selectedEntity] : [],
                 owner: user.user._id,
                 collaborator: [],
             })