+++ src/pages/RentalManagement/RentalManagementDetailsPage.tsx
@@ -11,7 +11,10 @@ import DetailsPage from "../../components/Shared/DetailsPage";
 import { useData } from "../../StateProvider/Provider";
 import CommonSkeleton from "../../components/Helpers/CommonSkeleton";
 import { CustomToastContext } from "../../StateProvider/CustomToastContext/CustomToastContext";
-import { getUniqueCurrencies, gridLoadingTimeout, rentalManagement, defaultActivityShow, dateFormat, pricingCondition, generateUniqueId, translateDataToTree } from "../../constants/helpers";
+import {
+  getUniqueCurrencies, gridLoadingTimeout, rentalManagement, defaultActivityShow,
+  dateFormat, pricingCondition, generateUniqueId, treeToFlatArray
+} from "../../constants/helpers";
 import Steps from "./Steps";
 import AddExistingProductInventory from "./AddExistingProductInventory";
 import CustomAgGrid, { intialState, reducer } from "../../components/AgGridComponents/CustomAgGrid";
@@ -65,7 +68,7 @@ const RentalManagementDetailsPage = () => {
   const [headingLbl, setHeadingLbl] = useState("");
   const [loadingDetails, setLoadingDetails] = useState(true);
   const [isUpdating, setUpdating] = useState(false);
-  const [isProductEdit, setProductEdit] = useState(false);
+  const [isProductEdit, setIsProductEdit] = useState(false);
   const [recordToUpdate, setRecordToUpdate] = useState(null)
   const [rentalManagementData, setRentalManagementData] = useState(null);
   const [deleteData, setDeleteData] = useState(null);
@@ -309,6 +312,50 @@ const RentalManagementDetailsPage = () => {
   const [state, dispatch] = useReducer(reducer, intialState);
   const { dataRows, rowCount, loading, page, limit, pageSizes, search, filters, sorting, selectedRecords } = state;
 
+  //  This is copied method from helpers.ts as wee need some modification for this screen only
+  const translateDataToTreeForProducts = (data, parentProperty, childProperty, childrenPropertyToStore) => {
+    let parents = data.filter(value => value[parentProperty] == 'undefined' || value[parentProperty] == null)
+    let childrens = data.filter(value => value[parentProperty] !== 'undefined' && value[parentProperty] != null)
+
+    parents.forEach((current) => {
+      if (current.type === "Product" || current.type === "Package") {
+        current["qtyToDisplay"] = current?.qty ?? 0;
+        current["isValid"] = (!isNaN(current?.finalPrice) && current?.finalPrice !== 0)
+      }
+    })
+
+    let translator = (parents, childrens) => {
+      parents.forEach((parent) => {
+        childrens.forEach((current, index) => {
+          if (current.parent === parent[childProperty]) {
+            let temp = JSON.parse(JSON.stringify(childrens))
+            temp.splice(index, 1)
+            translator([current], temp)
+
+            //  Check validation for products in package - Start
+            current["qtyToDisplay"] = `${current.qty} x ${parent.qty} = ${current.qty * parent.qty}`
+
+            if (current?.finalPrice !== null && current?.finalPrice !== undefined && typeof current?.finalPrice !== "string" && current?.finalPrice !== 0) {
+              current["isValid"] = true;
+            } else {
+              current["isValid"] = parent["isValid"];
+            }
+            //  Check validation for products in package - End
+
+            if (typeof parent[childrenPropertyToStore] !== 'undefined') {
+              parent[childrenPropertyToStore].push(current)
+            } else {
+              parent[childrenPropertyToStore] = [current]
+            }
+          }
+        })
+      })
+    }
+    translator(parents, childrens)
+
+    return parents
+  }
+
 
   const fetchProductInventory = () => {
 
@@ -323,6 +370,8 @@ const RentalManagementDetailsPage = () => {
         ...u,
         id: u._id,
         productCategory: u.productCategory?.optionLabel,
+        isValidRecord: true,
+        qtyToDisplay: u.qty
         // package: u.hasOwnProperty("package") ? u.package.packageName : "",
         // packageId: u.hasOwnProperty("package") ? u.package._id : ""
       })));
@@ -348,7 +397,9 @@ const RentalManagementDetailsPage = () => {
         newData.push(rest)
       })
 
-      const newDataForReactTable = [...translateDataToTree(newData ? [...newData] : [], "parent", "treeId", "subRows")];
+      const newDataForReactTable = [...translateDataToTreeForProducts(newData ? [...newData] : [], "parent", "treeId", "subRows")];
+
+      // console.log(newDataForReactTable)
 
       dispatch({ type: "initialize", data: [], count: 0 })
       dispatch({ type: "initialize", data: [...orderBy(newDataForReactTable, ["order"], ["asc"])], count: newDataForReactTable.length });
@@ -383,8 +434,6 @@ const RentalManagementDetailsPage = () => {
     })
 
     extractedProducts = [...products, ...packages, ...modifiedPkgProducts]
-
-    console.log(extractedProducts);
 
     return extractedProducts
   }
@@ -510,9 +559,10 @@ const RentalManagementDetailsPage = () => {
 
 
   const handleClick = (rowData) => {
-    setProductEdit(true)
+    setIsProductEdit(true)
     setRecordToUpdate(rowData)
   }
+
 
   const columns = [
     {
@@ -570,15 +620,10 @@ const RentalManagementDetailsPage = () => {
       )
     },
     {
-      accessor: 'qty',
+      accessor: 'qtyToDisplay',
       Header: 'Quantity',
       Cell: ({ row }) => (
-        row.original.qty ? <p className="text-truncate">
-          {/* {rowData.type === "productInPackage"
-              ? rowData.pkgQty !== 0 ? `${rowData.pkgQty} * ${rowData.qty} = ${rowData.totalQty}` : rowData.qty
-              : rowData.qty} */}
-          {row.original.qty}
-        </p> : <NoDataCell />
+        row.original.qtyToDisplay ? <p>{row.original.qtyToDisplay}</p> : <NoDataCell />
       )
     },
     {
@@ -599,7 +644,7 @@ const RentalManagementDetailsPage = () => {
       accessor: 'price',
       Header: `Price (${currencySymbol})`,
       Cell: ({ row }) => (
-        <p>{row.original.price ? row.original.price : "0"} </p>
+        row.original.price ? <p>{row.original.price}</p> : <NoDataCell />
       ),
       Footer: info => {
         const total = React.useMemo(
@@ -615,14 +660,14 @@ const RentalManagementDetailsPage = () => {
       accessor: 'discount',
       Header: 'Discount (%)',
       Cell: ({ row }) => (
-        <p>{row.original.discount === 0 ? "0" : row.original.discount}</p>
+        row.original.discount === 0 ? <NoDataCell /> : <p>{row.original.discount}</p>
       )
     },
     {
       accessor: 'finalPrice',
       Header: `Final Price (${currencySymbol})`,
       Cell: ({ row }) => (
-        <p>{row.original.finalPrice}</p>
+        row.original.finalPrice === 0 ? <NoDataCell /> : <p>{row.original.finalPrice}</p>
       ),
       Footer: info => {
         const total = React.useMemo(
@@ -730,8 +775,6 @@ const RentalManagementDetailsPage = () => {
   const handleBulkEditData = (values: any) => {
     let updatedArr = selectedProducts.filter(d => d.type !== "productInPackage")
 
-    console.log(updatedArr)
-
     updatedArr = updatedArr.map(d => ({
       "id": d.id,
       "type": d?.type.toLowerCase(),
@@ -752,7 +795,7 @@ const RentalManagementDetailsPage = () => {
     axiosInstance().put(`${rentalManagement.rentalManagementApi}/${id}/products-packages`, { "productsPackages": updatedArr })
       .then(() => {
         setUpdating(false)
-        setProductEdit(false)
+        setIsProductEdit(false)
         fetchProductInventory()
 
       }).catch((error) => {
@@ -764,19 +807,64 @@ const RentalManagementDetailsPage = () => {
 
   const handleSingleEdit = async (values: any) => {
     setUpdating(true)
-    const { subRows, ...rest } = values;
+    const { subRows, isValid, qtyToDisplay, ...rest } = values;
 
     rest.type = camelCase(rest.type)
     axiosInstance().put(`${rentalManagement.rentalManagementApi}/${id}/products-packages/updateOne`, rest)
       .then(() => {
         setUpdating(false)
-        setProductEdit(false)
+        setIsProductEdit(false)
         fetchProductInventory()
       }).catch((error) => {
         setUpdating(false)
         toastConfig.setToastConfig(error)
       });
   }
+
+  // const checkIsValidRecord = (rowData) => {
+  //   if (rowData?.type === "Product") {
+  //     return (!isNaN(rowData?.finalPrice) && rowData?.finalPrice !== 0)
+  //   }
+  //   else if (rowData?.type === "Package") {
+  //     return true;
+  //   }
+  //   else {
+  //     const parentRecord = dataRows.find(f => f.treeId === rowData?.parent)
+  //     if (parentRecord) {
+  //       return (!isNaN(parentRecord?.finalPrice) && parentRecord?.finalPrice !== 0)
+  //     }
+  //   }
+  // }
+
+  // const restrictToGoNextStep = () => {
+  //   let isValid = false;
+
+  //   dataRows.forEach(rowData => {
+  //     if (rowData?.type === "Product") {
+  //       if (isNaN(rowData?.finalPrice) || rowData?.finalPrice === 0) {
+  //         isValid = false;
+  //         return;
+  //       }
+  //     }
+  //     else if (rowData?.type === "Package") {
+  //       return true;
+  //     }
+  //     else {
+  //       const parentRecord = dataRows.find(f => f.treeId === rowData?.parent)
+  //       if (parentRecord) {
+  //         return (!isNaN(parentRecord?.finalPrice) && parentRecord?.finalPrice !== 0)
+  //       }
+  //     }
+  //   });
+
+  //   return isValid;
+
+  //   if (productInventory.length > 0) {
+  //     console.log(productInventory);
+  //     // !(productInventory.length > 0 ? (productInventory.filter(f => f.type !== "productInPackage").some(f => !f?.hasOwnProperty("finalPrice") || isNaN(f?.finalPrice) || f?.finalPrice === 0) ? false : true) : false)
+  //   }
+  //   return true;
+  // }
 
   return (
     <>
@@ -947,7 +1035,7 @@ const RentalManagementDetailsPage = () => {
                   <Paper>
                     <Steps
                       className={styles.steps_box}
-                      isNextStep={!(productInventory.length > 0 ? (productInventory.filter(f => f.type !== "productInPackage").some(f => !f?.hasOwnProperty("finalPrice") || isNaN(f?.finalPrice) || f?.finalPrice === 0) ? false : true) : false)}
+                      isNextStep={treeToFlatArray(dataRows, "subRows").some(f => f.isValid === false)}
                       nextStep={nextStep}
                       steps={rentalProcessSteps.slice(0, 5)}
                       currentStep={currentStep}
@@ -987,7 +1075,7 @@ const RentalManagementDetailsPage = () => {
                                   color="primary"
                                   size="small"
                                   disabled={!Boolean(selectedProducts && selectedProducts.length)}
-                                  onClick={() => setProductEdit(true)}
+                                  onClick={() => setIsProductEdit(true)}
                                 >
                                   Bulk Edit
                                 </Button>
@@ -1062,7 +1150,7 @@ const RentalManagementDetailsPage = () => {
                                 data={dataRows}
                                 rowStyle={(rowData) => ({
                                   color: "black",
-                                  backgroundColor: rowData?.type?.includes("roduct") && (isNaN(rowData?.finalPrice) || rowData?.finalPrice === 0) ? "#EFCCCC" : "white"
+                                  backgroundColor: rowData.isValid ? "white" : "#EFCCCC"
                                 })}
                                 onSelect={setSelectedProducts}
                                 childrenProperty="subRows"
@@ -1095,7 +1183,7 @@ const RentalManagementDetailsPage = () => {
                                   }
                                 }}
                               // onRowClick={(rowData) => {
-                              //   setProductEdit(true)
+                              //   setIsProductEdit(true)
                               //   setRecordToUpdate(rowData)
                               // }}
                               /> */}
@@ -1138,7 +1226,7 @@ const RentalManagementDetailsPage = () => {
                             { field: "description", headerName: "Description", show: true, disabled: true },
                             { field: "qty", headerName: "Quantity", show: true, disabled: true },
                             { field: "uom", headerName: "Unit of Measure", show: true, disabled: true },
-                            { field: "amount", headerName: `Amount ${currencySymbol}`, show: true, disabled: true },
+                            { field: "amount", headerName: `Amount (${currencySymbol})`, show: true, disabled: true },
                           ]}
                           dataRows={additionalCost}
                           frameworkComponents={rentalJobFrameworkComponents}
@@ -1360,7 +1448,7 @@ const RentalManagementDetailsPage = () => {
       {Boolean(packageForProducts)
         && <PackageProductsDialog
           rentalId={id}
-          packageId={packageForProducts?._id}
+          packageId={packageForProducts?.id}
           products={packageForProducts?.products.map(p => p.id)}
           onClose={() => setPackageForProducts(null)}
           rentalApi={rentalManagement.rentalManagementApi}
@@ -1377,7 +1465,7 @@ const RentalManagementDetailsPage = () => {
           endDate={rentalManagementData.rentalEndDate}
           isSaving={isUpdating}
           onClose={() => {
-            setProductEdit(false)
+            setIsProductEdit(false)
             setRecordToUpdate(null)
           }}
           submitBulkEdit={recordToUpdate ? handleSingleEdit : handleBulkEditData}